Removed function dependencies for module autonomy
A few functions from the function library now have access to a local function or tag under aestd.private instead of calling another AESTD function or an AESTD tag that would not be available if only some modules are loaded in a world.

Added a module display list in the minecraft:aestd function.

Added aestd1:entity/shoot_fireball.

Removed the player loot table which is no longer used by the function library.